Fix another race condition with status updating
Changed to not query the unit state on PropertiesChanged event but use the value received in the event instead.
This fixes that When restarting the tails-server-sftp.service, the PropertiesChanged event for state='inactive'
caused querying, but at the time the query was processed, the service was already started again, so it returned 'active' and the status never changed to 'inactive'.