Fix yet another race condition with status updating
Changed to add status updates to the main loop even if the status is the same as the old status.
Fixes that if a status changes so quickly forth and back that the main loop does not process the state change in between, then the old status would still be the same as in the beginning and the change back would not be processed.