XXX: add shitty Tails Server service option for starting the service *after* publishing the hidden service.
This is useful when the configuration needs the .onion, like
VirtualHost in prosody.

This implementation is breaks the status updating since it's currently
hardcoded (especially ServiceStatus's get_status_from_substates())
which type of event that has "priority" for setting the user visible
status. So, when the order is changed by this option, the dbus systemd
status monitoring will say that the process is "Stopped" while it is
announcing the onion.