Fix logic.
Previously we did:

    if diff_without_pot_creation_date ... && \
       diff_pot_only_line_comment_change ... ; then ...

but this won't work: the only time both can be true at the same time
is if both diffs are truly empty. E.g. if the header date is
different, it implies that there is a difference that is not a
comment, so the second operand will fail.

We have to check for both of these properties at the same time, not
separately in isolation.