Test suite: make Sikuli attempt to find replacements on FindFailed.
... by employing fuzz, or "lowering the similarity factor". The
replacements (if found) are saved among the artifacts, and serves as
potential drop-in-replacements for outdated images). The main use case
for this is when the font configuration in Tails changes, which
normally invalidates a large part of our images given that our default
high similarity factor.

We also add the `--fuzzy-image-matching` where the replacements are
used in case of FindFailed, so the tests can proceed beyond the first
FindFailed. The idea is that a full test suite run will produce
replacements for potentially *all* outdated images.