Tor Launcher: provide Tor's default Meek bridges.
It is not clear we want to do this, but for now we enable them as a
workaround:

In current Tor Launcher (whatever version shipped with Tor Browser
8.0a1), if no default bridges are configured the whole group of radio
buttons that provide the Moat option ("Request another bridge") is
hidden (with only the text box for manual entry available). So let's
provide some to enable Moat.

Refs: #8243