Reduce the Chutney Tor network to its minimal size.
The old values were picked pretty much arbitrarily, but with some
effort to make it seem realistic by enabling e.g. *some* circuit
variety. If we want that, we should specify better the aspects of the
real network that we want to keep. Until then, let's just make it as
small as possible for performance reasons; lately (especially since
tor 0.3.2.x) we've seen increased CPU and memory usage by each tor
process, to the point where it is noticeable on some systems we run it
on. Let's try to be nicer to them!

Refs: #15211