Specify IUK format version 2.
In our quest for (almost) endless incremental upgrades we'll end up
with larger IUKs than before. It turns out Tails Upgrader's memory
requirement of 2x the size of the IUK becomes a problem on systems
with 2 GB of memory, so we want to decrease the factor to 1x.

We achieve this by making the IUK mountable, so we can freely copy
contents from there onto the system partition without the need to
extract it *and* bundled archives, as was needed with the previous
format.

Will-fix: #6876
Refs: #11131