Test suite: Temoorary fix bug in APT hooks introduced by ASP.
At the moment, APT waits for the DPKG post install hook notification to
be clicked on before returning, which is a bug that should be fixed with
ticket #15382. Meanwhile we have to run APT install in the background
and check when the package is installed to go on with the feature.

Refs: #14596, #15382